Remove notice on customer_data_groups page
While testing the customer_data_groups.php page for an unrelated
reason, noticed that there was a notice in the admin hook.  This
removes that notice and should be easier to manage.

I'm unclear on whether the notice was something that I introduced
with a bad edit or a problem from PWA that I didn't notice before.
The logic said to show if one of three pages with an unset action
or if deleting.  My working hypothesis is that it was supposed to
be either the reviews page should show it if the action was empty
or delete or that all three pages were to show it for empty or
delete actions.

This flips the logic to check if it is one of the three pages and
$table_definition['columns'] is set.  So it won't run on the
customer_data_groups page nor will it attempt to run on any page
without a table_definition.